Move App into components; Remove Darkmode component
