+++ src/components/App.vue
@@ -5,9 +5,9 @@
 </template>
 
 <script>
-import Header from './components/Header'
-import Footer from './components/Footer'
-import GoogleForm from './components/GoogleForm'
+import Header from './Header'
+import Footer from './Footer'
+import GoogleForm from './GoogleForm'
 
 export default {
   name: 'App',
@@ -29,14 +29,3 @@ export default {
   }
 }
 </script>
-
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-
-  margin-top: 60px;
-}
-</style>
